fix(Tabulation2): fix data type

+++ tabulation/Tabulation2.java
@@ -1,7 +1,7 @@
 public class Tabulation2 {
     public static void main(String[] args) {
         for (int i = 1; i <= 50; i++) {
-            int power = 1;
+            long power = 1;
 
             for (int j = 1; j <= i; j++) {
                 power *= 2;
